Fixed filters bugs
